Backend/gl_predict-service/app_GL/requirements.txt
fastapi
uvicorn
pymysql
sqlalchemy
numpy
pandas
python-logstash
xgboost
joblib
scikit-learn
